requirements.txt
dnspython==2.6.1
colorama==0.4.4
